all the stuff I did yesterday
The heads up display should work just fine if you plug in the values with the set functions.

Start screen only goes to the audio menu and back the main menu.

graphics.gd is totally empty. I was going to work on it, but it's honestly kind of low priority.

HUD script now extends to CanvasLayer and is on the Root Node. It'd be a pain in the ass to fetch otherwise.